pull Corla CountyPoolsFromStyles out of CountyContestBuilder add CountyContestData for CountyAudit resampleAndRun() -> resampleAndSaveResults() move writingMvrs from runAuditRound to startNewRound
