+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/core/TestContest.kt
@@ -281,7 +281,7 @@ class TestContest {
 
         val expectedShow = """Contest 'testContestInfo' (0) PLURALITY voteForN=1 votes={1=108, 0=100, 2=0} undervotes=1, voteForN=1
    winners=[1] Nc=211 Nphantoms=2 Nu=1 sumVotes=208
-   1/0 votes=108/100 diff=8 (w-l)/w =0.0741 Npop=211 dilutedMargin=3.7915% recountMargin=7.4074% 
+   1/0 votes=108/100 diff=8 (w-l)/w =0.0741 Npop=211 minMargin=3.7915% recountMargin=7.4074% 
 
    0 'cand0': votes=100 
    1 'cand1': votes=108  (winner)
